requirements.txt
Flask==3.0.3
APScheduler==3.10.4
python-dateutil==2.9.0.post0
requests==2.32.3
